save env vars to file more safely
Before we exported environment variables to and `env` file using the command `env > jenkins-env`. This works until you have spaces in the value of the exported variables.

With an env variable like `ghprbPullDescription=Github pull request` the whole thing will break when reading the `jenkins-env` file back in. We have seen [this job](https://ci.centos.org/job/devtools-fabric8-ui/1493/console) for example where we've encountered this error:

```
++ ghprbPullDescription=GitHub
++ pull request
jenkins-env.clean: line 14: pull: command not found
```

This PR fixes the way we save environment variables by using `declare -p` instead of plain `env`. A variable like the above will be stored as `declare -x ghprbPullDescription="GitHub pull 111 request"` and can safely be read in using `source`.

For more information on this type `help declare` in a terminal or look at  [this post](https://unix.stackexchange.com/a/284397/124509).

All consumers of the `jenkins-env` file should be fine if they use `source` to read the values back in.