Please increase timeout for fabric8-services/wit coverage tests (#475)
* update coverage timeout for fabric8-wit to 90 minutes

Jenkins coverage builds for WIT have started failing with timeout errors.  
The current average 55-60 minutes, and the last build failed as it was nearing the end.
This patch adds enough headroom so this shouldn't be an issue.

* Increase timeout for fabric-8/wit coverage tests

Coverage for fabric8-wit is estimated to take about 2.5 hours currently.
This patch increases the timeout to 3 hours.