Block commits that remove/rename jobs
If a commit tries to remove or rename a job in the devtools view in
ci.centos.org, and the job already exists the test will fail.

The rationale behind this is that jobs are not automatically removed
from ci.centos.org when removed from the devtools index, therefore this
additional step forces the admin to manually remove the jobs from
jenkins before accepting the commit.