WIP: Issue #580: Use python's pipes.quote to quote vars
There are many statements build scripts that dump the environment like
this: `env > jenkins-env`. All of these have been replace by a python
snippet that iterates through all the environment variables and dumps
them safely using
[pipes.quote](https://docs.python.org/2/library/pipes.html#pipes.quote)

The problem that this does not solve is multiline env vars. Many OSIO
scripts will do a `grep <field>` first, dump to a new file, and then
source that file. This is inconvenient as it will discard all but the
first line of the multiline env vars.