Simplify this script
Before it was using the jenkins-jobs internal API to get the list of
jobs. It also supported multiple ways of fetching the list of jobs to
compare it with.

All this complexity has been removed. Now the list of jobs is read from
STDIN and it always uses the CentOS CI Jenkins as the list of jobs to
compare it with.