Improve PR check (#778)
* Simplify this script

Before it was using the jenkins-jobs internal API to get the list of
jobs. It also supported multiple ways of fetching the list of jobs to
compare it with.

All this complexity has been removed. Now the list of jobs is read from
STDIN and it always uses the CentOS CI Jenkins as the list of jobs to
compare it with.

* Improve the PR check

- USE_LOCAL to run the PR job in the same folder and NOT remove the temp
  files.
- Support overriding of the Jenkins Jobs path.
- Improved way to handle removing the temp files (using a trap).
- Make sure the admin list is in alphabetical order.
- Use the newly simplified jenkins-jobs-diff.py

* sort the admin list

* Remove alphabetical check since we will move to org auth

* Revert "sort the admin list"

This reverts commit 584a7c055c090d0c6c54feb1c9d03569e291563f.

* do not hardcode path

* typo

* cannot use quotes in this case